+++ packages/ethereum-contracts/contracts/interfaces/superfluid/Definitions.sol
@@ -231,17 +231,22 @@ library BatchOperation {
      */
     uint32 constant internal OPERATION_TYPE_SUPERFLUID_CALL_APP_ACTION = 2 + 200;
     /**
-     * @dev DMZForwarder.forwardCall batch operation type
+     * @dev SimpleForwarder.forwardCall batch operation type
      *
      * Call spec:
      * forwardCall(
      *     target,
      *     data
      * )
+     * NOTE: This operation allows to make arbitrary calls to arbitrary targets.
+     * The calls are routed through a dedicated utility contract `SimpleForwarder`.
+     * This is important because the host contract has privileged access to other framework contracts,
+     * SuperTokens, SuperApps etc.
+     * Allowing arbitrary calls to arbitrary targets with the host as sender would thus be unsafe.
      */
     uint32 constant internal OPERATION_TYPE_SIMPLE_FORWARD_CALL = 1 + 300;
     /**
-     * @dev DMZForwarder.forward2771Call batch operation type
+     * @dev ERC2771Forwarder.forward2771Call batch operation type
      *
      * Call spec:
      * forward2771Call(
@@ -250,17 +255,17 @@ library BatchOperation {
      *     data
      * )
      *
-     * NOTE: In the context of this operation, the `DZMForwarder` contract acts as the
+     * NOTE: In the context of this operation, the `ERC2771Forwarder` contract acts as the
      * _trusted forwarder_ which must be trusted by the _recipient contract_ (operation target).
-     * It shall do so by dynamically looking up the DMZForwarder used by the host, like this:
+     * It shall do so by dynamically looking up the ERC2771Forwarder used by the host, like this:
      *
      * function isTrustedForwarder(address forwarder) public view returns(bool) {
-     *     return forwarder == address(host.DMZ_FORWARDER());
+     *     return forwarder == address(host.getERC2771Forwarder());
      * }
      *
      * If used in the context of a `forwardBatchCall`, we effectively have a chaining/nesting
      * of ERC-2771 calls where the host acts as _recipient contract_ of the enveloping 2771 call
-     * and the DMZForwarder acts as the _trusted forwarder_ of the nested 2771 call(s).
+     * and the ERC2771Forwarder acts as the _trusted forwarder_ of the nested 2771 call(s).
      * That's why `msgSender` could be either the actual `msg.sender` (if using `batchCall`)
      * or the relayed sender address (if using `forwardBatchCall`).
      */